Render helpful error for graphiti invalid request
Graphiti is introducing a new error type which will capture validation
around problems with the payload and query parameters provided in a
request.  This commit introduces a serializer and request handler for
the graphiti exception type, and auto-registers the handler so that
clients who upgrade will get the new behavior out of the box.

This also restructures the directories a bit more to bucket classes into
the type of error they are helping to manage.  The only externally
visible change is the move from
`GraphitiErrors::Serializers::Validation` to
`GraphitiErrors::Validation::Serializer`, which is referenced in
graphiti's railtie for 1.0.x (and beta versions).  To make this upgrade
easier, I am also declaring the old constant to equal the new one so
that upgrades to this won't break older consumers.